requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pytest
mypy

jupyter
jupyter_contrib_nbextensions
numpy
pandas
matplotlib
seaborn
sklearn
pandas_profiling
xgboost
plotly
